+++ zephyr-mcumgr/src/connection.rs
@@ -2,7 +2,7 @@ use std::{io::Cursor, sync::Mutex, time::Duration};
 
 use crate::{
     commands::{ErrResponse, ErrResponseV2, McuMgrCommand},
-    smp_errors::DeviceError,
+    smp_errors::{DeviceError, MCUmgrErr},
     transport::{ReceiveError, SendError, Transport},
 };
 
@@ -46,6 +46,17 @@ pub enum ExecuteError {
     #[error("Device returned error code: {0}")]
     #[diagnostic(code(zephyr_mcumgr::connection::execute::device_error))]
     ErrorResponse(DeviceError),
+}
+
+impl ExecuteError {
+    /// Checks if the device reported the command as unsupported
+    pub fn command_not_supported(&self) -> bool {
+        if let Self::ErrorResponse(DeviceError::V1 { rc, .. }) = self {
+            *rc == MCUmgrErr::MGMT_ERR_ENOTSUP as i32
+        } else {
+            false
+        }
+    }
 }
 
 impl Connection {
@@ -121,10 +132,12 @@ impl Connection {
         }
 
         if let Some(rc) = err.rc {
-            return Err(ExecuteError::ErrorResponse(DeviceError::V1 {
-                rc,
-                rsn: err.rsn,
-            }));
+            if rc != MCUmgrErr::MGMT_ERR_EOK as i32 {
+                return Err(ExecuteError::ErrorResponse(DeviceError::V1 {
+                    rc,
+                    rsn: err.rsn,
+                }));
+            }
         }
 
         let decoded_response: R::Response = ciborium::from_reader(Cursor::new(response))