+++ CHANGELOG.md
@@ -5,6 +5,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.7.1] - xxxx-xx-xx
+
+### Changes
+
+- Increase default timeout to `10000 ms`.
+
+### Fixes
+
+- Firmware upload to MCUboot recovery mode failed with `MGMT_ERR_EOK`
+
 ## [0.7.0] - 2026-01-24
 
 ### Breaking Changes